app/requirements.txt
Django
djangorestframework
djangorestframework-jwt
psycopg2-binary
gunicorn
django-cors-headers

